requirements.txt
google-generativeai
playwright
playwright-stealth
requests
beautifulsoup4
pypdf
python-dateutil
reportlab
setuptools 